Drop `getAuthCodeInteractive`
It was always a bit of an ugly hack for convenience. The `login` executable is much nicer than the Spotipy approach.
