+++ login/Main.hs
@@ -6,15 +6,18 @@ import Control.Monad
 import Data.Function
 import Data.Functor
 import Data.Proxy
+import Data.Set (Set)
 import Data.Text qualified as T
 import Data.Text.IO qualified as T
 import Lucid (Html, ToHtml (toHtml), div_, pre_)
 import Network.HTTP.Client.TLS (newTlsManager)
 import Network.Wai.Handler.Warp (Port, defaultSettings, runSettings, setBeforeMainLoop, setPort)
 import Options.Applicative
-import Servant (Get, QueryParam', Required, Strict, serve, type (:>))
+import Servant (Get, QueryParam', Required, Strict, allLinks, linkURI, serve, type (:<|>) ((:<|>)), type (:>))
+import Servant.Client (baseUrlHost)
 import Servant.HTML.Lucid (HTML)
 import Spotify
+import Spotify.Servant qualified
 import Web.Browser (openBrowser)
 
 main :: IO ()
@@ -45,3 +48,23 @@ run port clientId clientSecret = do
   where
     scopes = Just allScopes
     redirect = URL $ "http://127.0.0.1:" <> T.pack (show port)
+
+authorizeUrl :: ClientId -> URL -> Maybe (Set Scope) -> URL
+authorizeUrl clientId redirectURI scopes =
+    URL $
+        "https://"
+            <> T.pack
+                ( baseUrlHost accountsBase
+                    <> "/"
+                    <> show (linkURI link)
+                )
+  where
+    link =
+        link0
+            clientId
+            "code"
+            redirectURI
+            Nothing
+            (ScopeSet <$> scopes)
+            Nothing
+    _ :<|> _ :<|> link0 = allLinks $ Proxy @Spotify.Servant.AccountsAPI
